Move the cascade logic out of the AccrualModule into the ContributionsHandler
