+++ src/main/java/uk/gov/homeoffice/digital/sas/balancecalculator/handlers/ContributionsHandler.java
@@ -1,5 +1,6 @@
 package uk.gov.homeoffice.digital.sas.balancecalculator.handlers;
 
+import static uk.gov.homeoffice.digital.sas.balancecalculator.constants.Constants.ACCRUALS_MAP_EMPTY;
 import static uk.gov.homeoffice.digital.sas.balancecalculator.constants.Constants.MISSING_ACCRUAL;
 import static uk.gov.homeoffice.digital.sas.balancecalculator.constants.Constants.NO_ACCRUALS_FOUND_FOR_TYPE;
 import static uk.gov.homeoffice.digital.sas.balancecalculator.constants.Constants.OPERATION_NOT_IMPLEMENTED;
@@ -10,6 +11,7 @@ import java.text.MessageFormat;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.SortedMap;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
@@ -64,16 +66,10 @@ public class ContributionsHandler {
           return false;
         }
 
-        updateAccrualContribution(timeEntry.getId(), contribution, accrual, action);
+        this.updateAccrualContribution(timeEntry.getId(), contribution, accrual, action);
       }
 
-
-      boolean result = module.applyTimeEntryToAccruals(timeEntry, action,
-          accruals, applicableAgreement);
-
-      if (!result) {
-        return false;
-      }
+      this.cascadeCumulativeTotal(accruals, applicableAgreement.getStartDate());
     }
     return true;
   }
@@ -100,4 +96,47 @@ public class ContributionsHandler {
     contributions.setTotal(total);
   }
 
+  void cascadeCumulativeTotal(
+      SortedMap<LocalDate, Accrual> accruals, LocalDate agreementStartDate) {
+
+    Optional<LocalDate> optional = accruals.keySet().stream().findFirst();
+    if (optional.isPresent()) {
+      LocalDate priorAccrualDate = optional.get();
+      Accrual priorAccrual = accruals.get(priorAccrualDate);
+
+      BigDecimal baseCumulativeTotal = BigDecimal.ZERO;
+      // if prior accrual is related to the same agreement then use its cumulative total
+      // as starting point; otherwise, start at 0
+      if (isPriorAccrualRelatedToTheSameAgreement(priorAccrualDate, agreementStartDate)) {
+        baseCumulativeTotal = priorAccrual.getCumulativeTotal();
+      }
+
+      // the first element is only used to calculate base cumulative total so shouldn't be included
+      List<Accrual> accrualsFromReferenceDate = accruals.values().stream().skip(1).toList();
+      this.updateSubsequentAccruals(accrualsFromReferenceDate, baseCumulativeTotal);
+    } else {
+      throw new IllegalArgumentException(ACCRUALS_MAP_EMPTY);
+    }
+  }
+
+  private boolean isPriorAccrualRelatedToTheSameAgreement(
+      LocalDate priorAccrualDate, LocalDate agreementStatDate) {
+    return !priorAccrualDate.isBefore(agreementStatDate);
+  }
+
+  void updateSubsequentAccruals(List<Accrual> accruals, BigDecimal priorCumulativeTotal) {
+
+    //update the cumulative total for referenceDate
+    accruals.get(0).setCumulativeTotal(
+        priorCumulativeTotal.add(accruals.get(0).getContributions().getTotal()));
+
+    //cascade through until end of agreement
+    for (int i = 1; i < accruals.size(); i++) {
+      BigDecimal priorTotal =
+          accruals.get(i - 1).getCumulativeTotal();
+      Accrual currentAccrual = accruals.get(i);
+      currentAccrual.setCumulativeTotal(
+          priorTotal.add(currentAccrual.getContributions().getTotal()));
+    }
+  }
 }
